Added Icons and Styled
I elected to add individual SVG Icons for use on these cards, but this can be altered later if we elect to use a full icon package or something. I "finished" styling the card, but I have some concerns about responsiveness that we should check back on before public deployment.